show short header on non-home pages
