+++ app/javascript/src/App/Layout/Header/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { Header } from './Header'
